move EQ definitions into definitions
